scorer/requirements.txt
python-arango==7.5.1
py_expression_eval==0.3.14
